Fixed tests to use updated BuildEngine.

+++ test/test_build_engine.py
@@ -3,6 +3,8 @@ import unittest
 import shutil
 import os
 import example_models
+
+import gillespy2
 from gillespy2.solvers.cpp.build.build_engine import BuildEngine
 
 
@@ -18,12 +20,16 @@ class TestBuildEngine(unittest.TestCase):
     ]
 
     def setUp(self) -> None:
+        # Disable the gillespy2 cache.
+        gillespy2.cache_enabled = False
+
         self.build_engine = BuildEngine()
-        self.tmp_dir = str(self.build_engine.temp_dir)
+        self.tmp_dir = self.build_engine.prepare(self.test_model, variable=False)
 
     def tearDown(self) -> None:
         self.build_engine.clean()
-        if os.path.exists(self.tmp_dir):
+
+        if self.tmp_dir is not None and os.path.exists(self.tmp_dir):
             shutil.rmtree(self.tmp_dir, ignore_errors=True)
 
     def test_default_layout(self):
@@ -32,15 +38,13 @@ class TestBuildEngine(unittest.TestCase):
           are subdirectories of the temp directory.
         This is to ensure that they get cleaned up when the temp directory is cleaned up.
         """
-        self.build_engine.prepare(self.test_model, variable=False)
-
         # This test should be attempted with each solver target.
         for solver_name in self.solver_names:
             with self.subTest(solver=solver_name):
                 self.build_engine.build_simulation(solver_name)
 
         template_dir = self.build_engine.template_dir
-        obj_dir = self.build_engine.cache_dir
+        obj_dir = self.build_engine.obj_dir
 
         self.assertTrue(str(template_dir.resolve()).startswith(str(Path(self.tmp_dir).resolve())),
                         "Template directory was not placed in temp directory when cache is disabled")
@@ -68,7 +72,6 @@ class TestBuildEngine(unittest.TestCase):
         """
         template_file = self.build_engine.template_dir.joinpath(
                         self.build_engine.template_definitions_name)
-        self.build_engine.prepare(self.test_model, variable=False)
         self.assertTrue(template_file.exists(), "Simulation template header file could not be found")
 
         # Test to ensure the contents of the output template_file contains ONLY macro definitions,
@@ -82,9 +85,8 @@ class TestBuildEngine(unittest.TestCase):
         Ensure that the default, cache-less output has expected behavior.
         Should result in an expected definitions file and an executable simulation.
         """
-        simulation_file = self.build_engine.temp_dir.joinpath(
+        simulation_file = self.build_engine.output_dir.joinpath(
             "Simulation.exe" if os.name == "nt" else "Simulation.out")
-        self.build_engine.prepare(self.test_model, variable=False)
 
         for solver_name in self.solver_names:
             with self.subTest(solver=solver_name):
